# swagger.yaml
openapi: "3.0.0"
info:
  title: Redstone Goats Order API
  description: API for creating and managing orders with UBL XML document generation. State diagram is available at https://ibb.co/21H16g4J.
  version: 1.0.0
servers:
  - url: https://h13-redstone-goats.vercel.app
...
paths:      
  /v1/user/register:
    post:
      summary: Register a new user.
      description: Registers a new user and returns the userId.
      tags:
        - "User Routes"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                nameFirst:
                  type: string
                nameLast:
                  type: string
      responses:
        '200': 
          description: Success
          content:
            application/json:
              schema: 
                type: object
                properties:
                  userId:
                    type: integer
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                    example: "Error Message Here"
  /v1/user/login:
    post:
      summary: Login an existing user
      description: Authenticate a user with email and password.
      tags:
        - "User Routes"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: integer
        '400':
          description: Bad request from client
          content:
            application/json:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                    example: "Error Message Here"
        '401':
          description: Token is unauthorized/invalid
          content:
            application/json:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                    example: "Error Message Here"
  /v1/user/logout:
    post:
      summary: Logout an existing user
      description: Logout a user by blacklisting the JWT token.
      tags:
        - "User Routes"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                Token:
                  type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Bad request from client
          content:
            application/json:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                    example: "Error Message Here"
        '401':
          description: Token is unauthorized/invalid
          content:
            application/json:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                    example: "Error Message Here"
  /v1/user/details:
    get:
      summary: Gets a user's details
      description: Retrieve the details about a specific user
      tags:
        - "User Routes"
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
          description: ID of the user to retrieve details for
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: integer
                  name:
                    type: string
                  email:
                    type: string
                  numSuccessfulLogins: 
                    type: integer
                  numFailedPasswordsSinceLastLogin:
                    type: integer
        '400':
          description: Bad request from client
          content:
            application/json:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                    example: "Error Message Here"
        '401':
          description: Token is unauthorized/invalid
          content:
            application/json:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                    example: "Error Message Here"
    put:
      summary: Updates a user's details
      description: Change the current user details
      tags:
        - "User Routes"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
                email:
                  type: string
                nameFirst:
                  type: string
                nameLast:
                  type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Bad request from client
          content:
            application/json:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                    example: "Error Message Here"
        '401':
          description: Token is unauthorized/invalid
          content:
            application/json:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                    example: "Error Message Here"
  /v1/order/create:
    post:
      summary: Creates an order and produces a UBL XML document.
      description: Creates an order with items, user details, billing details, and delivery instructions.
      tags:
        - "Buyer Routes"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderRequest"
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                type: object
                properties:
                  orderId:
                    type: integer
        '400':
          description: Bad request from client e.g invalid bank details, invalid order date
          content:
            application/json:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                    example: "Error Message Here"
        '401':
          description: Token is unauthorized/invalid or a different name is registered to the userId.
          content:
            application/json:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                    example: "Error Message Here"
...
  /v1/{userId}/order/{orderId}/cancel:
    put:
      summary: Cancel an order before dispatch.
      description: Cancels an existing order and adds it to the user's cancel order history.
      tags:
        - "Buyer Routes"
        - "Seller Routes"
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Order canceled
          content:
            application/json:
              schema:
                type: object
                properties:
                  cancelled:
                    type: boolean
                  reason:
                    type: string
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                    example: "Error Message Here"
        '401':
          description: token invalid or unauthorised
          content:
            application/json:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                    example: "Error Message Here"
        '404':
          description: Order not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                    example: "Error Message Here"
...
  /v1/{userId}/order/history:
    post:
      summary: Retrieve user order history.
      description: Retrieves the order history of the specific buyer.
      tags:
        - "Buyer Routes"
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Order history retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderHistory"
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                    example: "Error Message Here"
        '401':
          description: Token invalid or unauthorised. 
          content:
            application/json:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                    example: "Error Message Here"
...
  /v1/order/{userId}/sales:
    get:
      summary: Send sales analytics data to the seller.
      description: Sends analytical data to the product distributor, with different options for data format.
      tags:
        - "Seller Routes"
      parameters:
        - name: CSV
          in: query
          schema:
            type: boolean
            default: true
        - name: JSON
          in: query
          schema:
            type: boolean
            default: true
        - name: PDF
          in: query
          schema:
            type: boolean
            default: true
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Sales data sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesData"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                    example: "Error Message Here"
        '401':
          description: Token unauthorised or invalid
          content:
            application/json:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                    example: "Error Message Here"
  /v1/user/item/add:
    post:
      summary: Add items which can be purchased by other users.
      description: Allows a user (which can be both a buyer and/or a seller) to put items up for sale.
      tags:
        - "Seller Routes"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: "#/components/schemas/ItemV2"
      responses:
        '200':
          description: Items added successfully
        '400':
          description: Bad request from client e.g invalid seller name, phone number, item price.
          content:
            application/json:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                    example: "Error Message Here"
  /v1/order/{orderId}/details:
    get:
      summary: Finds and returns order details given an order Id.
      description: Finds and returns order details given an order Id.
      tags:
        - "Seller Routes"
        - "Buyer Routes"
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Order retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderRequestV2"
        '400':
          description: Bad request from client e.g invalid orderId.
          content:
            application/json:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                    example: "Error Message Here"
  /v1/item/{itemId}/details:
    get:
      summary: Finds and returns item details given an item Id.
      description: Finds and returns item details given an item Id.
      tags:
        - "Buyer Routes"
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Item retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemV2"
        '400':
          description: Bad request from client e.g invalid itemId.
          content:
            application/json:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                    example: "Error Message Here"
  /v1/{userId}/item/all/details:
    get:
      summary: Finds and returns all items that exist.
      description: Finds and returns all items that exist.
      tags:
        - "Buyer Routes"
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Items retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: 
                      $ref: "#/components/schemas/ItemV2"
        '400':
          description: Bad request from client e.g invalid itemId.
          content:
            application/json:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                    example: "Error Message Here"
  /v1/order/{orderId}/XML/render:
    get:
      summary: Finds and returns an XML string given an orderId.
      description: Finds and returns the corresponding XML string for an order.
      tags:
        - "Buyer Routes"
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: XML string retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  xmlDocument:
                    type: string
        '400':
          description: Bad request from client e.g invalid orderId.
          content:
            application/json:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                    example: "Error Message Here"
  /v2/order/create:
    post:
      summary: Creates an order and produces a UBL XML document.
      description: Version 2 of the order creation route. Request body includes
        extra fields such as the tax amount, taxTotal, currency, paymentAccountId, paymentAccountName, financialInstitutionBranch. In addition extra phone number and email fields are required for the sellers and buyer information.
      tags:
        - "Version 2 Routes"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderRequestV2"
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                type: object
                properties:
                  orderId:
                    type: integer
        '400':
          description: Bad request from client e.g invalid bank details, invalid order date
          content:
            application/json:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                    example: "Error Message Here"
        '401':
          description: Token is unauthorized/invalid or a different name is registered to the userId.
          content:
            application/json:
              schema:
                type: object
                properties:
                  Error:
                    type: string
                    example: "Error Message Here"
components:
  schemas:
    OrderRequest:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
        quantities:
          type: array
          items:
            type: number
        buyer:
          $ref: "#/components/schemas/UserSimple"
        billingDetails:
          $ref: "#/components/schemas/BillingDetailsParam"
        delivery:
          $ref: "#/components/schemas/DeliveryInstructions"
        lastEdited:
          type: string
          format: date-time
        status:
          type: string
          enum:
            - pending
            - confirmed
            - cancelled
        totalPrice:
          type: number
        createdAt:
          type: string
          format: date-time
    OrderRequestV2:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ItemV2"
        quantities:
          type: array
          items:
            type: number
        buyer:
          $ref: "#/components/schemas/UserSimpleV2"
        billingDetails:
          $ref: "#/components/schemas/BillingDetailsParam"
        delivery:
          $ref: "#/components/schemas/DeliveryInstructions"
        lastEdited:
          type: string
          format: date-time
        status:
          type: string
          enum:
            - pending
            - confirmed
            - cancelled
        totalPrice:
          type: number
        taxAmount:
          type: number
        taxTotal:
          type: number
        currency:
          type: string
        paymentAccountId:
          type: string
        paymentAccountName:
          type: string
        financialInstitutionBranch:
          type: string
        createdAt:
          type: string
          format: date-time
    UserSimple:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        streetName:
          type: string
        cityName:
          type: string
        postalZone:
          type: string
        cbcCode:
          type: string
...
    BillingDetailsParam:
      type: object
      properties:
        creditCardNumber:
          type: string
        CVV:
          type: integer
          format: int32
        expiryDate:
          type: string
          format: date
...
    Item:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        price:
          type: number
        description:
          type: string
        seller:
          type: object
          $ref: "#/components/schemas/UserSimple"
    ItemV2:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        price:
          type: number
        description:
          type: string
        seller:
          type: object
          $ref: "#/components/schemas/UserSimpleV2"      
...
    OrderHistory:
      type: object
      properties:
        successfulOrders:
          type: array
          items: 
            $ref: "#/components/schemas/OrderRequest"
        cancelledOrders:
          type: array
          items: 
            $ref: "#/components/schemas/OrderRequest"
...
    SalesData:
      type: object
      properties:
        sales:
          type: array
          items:
            type: object
            properties:
              id:
                type: number
              name:
                type: string
              description:
                type: string
              price:
                type: number
              amountSold:
                type: integer
        CSVurl:
          type: string
        PDFurl:
          type: string